major/v2-architecture: finalized migration
